requirements.txt
loguru>=0.7.0
playwright>=1.40.0
pyotp>=2.9.0
cryptography>=42.0.0
opencv-python>=4.8.0
pytest-asyncio>=0.21.0
